+++ utils.py
@@ -1,0 +1,32 @@
+
+# Import datasets, classifiers and performance metrics
+from sklearn import svm,datasets
+from sklearn.model_selection import train_test_split
+
+#read gigits
+def read_digits():
+    digits = datasets.load_digits()
+    x = digits.images
+    y = digits.target 
+    return x,y
+
+# We will define utils here :
+def preprocess_data(data):
+    n_samples = len(data)
+    data = data.reshape((n_samples, -1))
+    return data
+
+# Split data into 50% train and 50% test subsets
+
+def split_data(X,y,test_size=0.5,random_state=1):
+    X_train, X_test, y_train, y_test = train_test_split(
+     X, y, test_size=test_size, shuffle=False
+    )
+    return X_train, X_test, y_train, y_test
+
+# Create a classifier: a support vector classifier
+def train_model(X, y, model_params,model_type = 'svm'):
+    if model_type == 'svm':
+        clf = svm.SVC(**model_params)
+    clf.fit(X, y)
+    return clf
